(Carlos) Atualização das classes para api Rest (EnderecoController, EnderecoService, Pedido, PedidoController, PedidoService, CategoriaController, CategoriaService)

+++ main/src/main/java/com/atividade/main/model/Pedido.java
@@ -32,7 +32,7 @@ public class Pedido {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int pedidoID;
+	private long pedidoID;
 
 	@OneToOne
 	@JoinColumn(name = "userID")
